requirements.txt
gradio
matplotlib
numpy
pandas
Pillow
scikit_learn
scipy
skimage
ultralytics
opencv-python